added post-processing scripts

+++ ImagePreSegmentor.py
@@ -10,17 +10,18 @@
 import SegmentationFunctions
 import utils
 from PIL import Image
-import os, glob, pickle
+import os
+import glob
+import pickle
 import numpy as np
 import cv2
 import imageio
 from pathlib import Path
-import scipy
 
 
 class ImagePreSegmentor:
 
-    def __init__(self, dir_to_process, dir_output, dir_model):
+    def __init__(self, dir_to_process, dir_output, dir_model, img_type, overwrite):
         self.dir_to_process = Path(dir_to_process)
         self.dir_model = Path(dir_model)
         # output paths
@@ -28,6 +29,8 @@ class ImagePreSegmentor:
         self.path_mask = self.path_output / "Mask"
         self.path_proba = self.path_output / "Proba"
         self.path_overlay = self.path_output / "Overlay"
+        self.image_type = img_type
+        self.overwrite = overwrite
 
     def prepare_workspace(self):
         """
@@ -38,14 +41,14 @@ class ImagePreSegmentor:
         self.path_proba.mkdir(parents=True, exist_ok=True)
         self.path_overlay.mkdir(parents=True, exist_ok=True)
 
-    def file_feed(self, img_type):
+    def file_feed(self):
         """
         Creates a list of paths to images that are to be processed
         :param img_type: a character string, the file extension, e.g. "JPG"
         :return: paths
         """
         # get all files and their paths
-        files = glob.glob(f'{self.dir_to_process}/*.{img_type}')
+        files = glob.glob(f'{self.dir_to_process}/*.{self.image_type}')
         return files
 
     def segment_image(self, img):
@@ -70,7 +73,7 @@ class ImagePreSegmentor:
 
         # threshold image on probabilities
         # TODO define a better threshold value
-        binary_mask = np.where(probabilities > 0.65, 1, 0)
+        binary_mask = np.where(probabilities > 0.5, 1, 0)
 
         # smooth binary mask
         # TODO optimize the kernel size
@@ -79,16 +82,16 @@ class ImagePreSegmentor:
         # TODO optimize the filters
         mask_filtered = utils.filter_objects_size(binary_mask, 150, "smaller")
         m_inv = cv2.bitwise_not(mask_filtered*255)
-        mask_filtered = utils.filter_objects_size(m_inv, 15, "smaller")
+        mask_filtered = utils.filter_objects_size(m_inv, 150, "smaller")
         mask_filtered = cv2.bitwise_not(mask_filtered)
 
         # create overlay
         M = mask_filtered.ravel()
         M = np.expand_dims(M, -1)
-        outmask = np.dot(M, np.array([[255, 0, 0, 75]]))
-        outmask = np.reshape(outmask, newshape=(img.shape[0], img.shape[1], 4))
-        outmask = outmask.astype("uint8")
-        mask = Image.fromarray(outmask, mode="RGBA")
+        out_mask = np.dot(M, np.array([[1, 0, 0, 0.33]]))
+        out_mask = np.reshape(out_mask, newshape=(img.shape[0], img.shape[1], 4))
+        out_mask = out_mask.astype("uint8")
+        mask = Image.fromarray(out_mask, mode="RGBA")
         img_ = Image.fromarray(img, mode="RGB")
         img_ = img_.convert("RGBA")
         img_.paste(mask, (0, 0), mask)
@@ -96,32 +99,33 @@ class ImagePreSegmentor:
 
         return probabilities, mask_filtered, overlay
 
-    def segment_images(self, img_type):
+    def segment_images(self):
         """
         Wrapper, processing all images
         :param img_type: a character string, the file extension, e.g. "JPG"
         """
         self.prepare_workspace()
-        files = self.file_feed(img_type)
+        files = self.file_feed()
 
         for file in files:
 
             # get file basename
-            basename = os.path.basename(file)
-            pngname = basename.replace("." + img_type, ".png")
-
-            img = imageio.imread(file)
-            proba, mask, overlay = self.segment_image(img)
+            base_name = os.path.basename(file)
+            png_name = base_name.replace("." + self.image_type, ".png")
 
             # output paths
-            proba_name = self.path_proba / basename
-            mask_name = self.path_mask / pngname
-            overlay_name = self.path_overlay / pngname
+            proba_name = self.path_proba / base_name
+            mask_name = self.path_mask / png_name
+            overlay_name = self.path_overlay / png_name
 
-            # print(overlay_name)
+            if not self.overwrite and os.path.exists(mask_name):
+                continue
+            else:
+                img = imageio.imread(file)
+                proba, mask, overlay = self.segment_image(img)
 
-            # save masks
-            imageio.imwrite(mask_name, mask)
-            imageio.imwrite(proba_name, proba)
-            imageio.imwrite(overlay_name, overlay)
+                # save masks
+                imageio.imwrite(mask_name, mask)
+                imageio.imwrite(proba_name, proba)
+                imageio.imwrite(overlay_name, overlay)
